fix(frontend): address management review feedback
